+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.1] - 2022-11-18
 ### Fixed
+- Docs.
 - Run junk added to poetry.
 - Run junk added to poetry.
 
@@ -18,8 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial project create.
 
-### Changed
-- Initial commit.
+### Fixed
+- Run junk added to poetry.
 
 [Unreleased]: https://github.com/OpenTransitTools/yaml-json-tools/compare/0.1.1...HEAD
 [0.1.1]: https://github.com/OpenTransitTools/yaml-json-tools/compare/0.1.0...0.1.1